+++ src/features/userPage/UserPage.tsx
@@ -1,6 +1,6 @@
 import { useSelector } from 'react-redux';
 
-import { RootState } from '../../app/store/store';
+import { RootState } from '../../app/store';
 
 const UserPage = () => {
   const aaa = useSelector((state: RootState) => state.user.name);
