+++ src/index.tsx
@@ -8,7 +8,7 @@ import { Toaster } from 'react-hot-toast';
 import { createRoot } from 'react-dom/client';
 
 import App from './App';
-import store from './app/store/store';
+import store from './app/store';
 
 const container = document.getElementById('root')!;
 const root = createRoot(container);
